moving stuff around (2)
